changed Integer to Double

+++ seg/src/main/java/com/application/dashboard/GraphGenerator.java
@@ -24,9 +24,9 @@ public class GraphGenerator {
     private String title;
     private String xAxisLabel;
     private String yAxisLabel;
-    private Map<LocalDateTime, Integer> data;
+    private Map<LocalDateTime, Double> data;
     private ChartFrame frame;
-    public GraphGenerator(String title, String xAxisLabel, String yAxisLabel, Map<LocalDateTime, Integer> data) {
+    public GraphGenerator(String title, String xAxisLabel, String yAxisLabel, Map<LocalDateTime, Double> data) {
         this.title = title;
         this.xAxisLabel = xAxisLabel;
         this.yAxisLabel = yAxisLabel;
@@ -49,7 +49,7 @@ public class GraphGenerator {
         return yAxisLabel;
     }
 
-    public Map<LocalDateTime, Integer> getData() {
+    public Map<LocalDateTime, Double> getData() {
         return data;
     }
 
@@ -67,7 +67,7 @@ public class GraphGenerator {
         // Create a time series collection and add data points to it
         TimeSeries series = new TimeSeries(yAxisLabel);
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
-        for (Map.Entry<LocalDateTime, Integer> entry : data.entrySet()) {
+        for (Map.Entry<LocalDateTime, Double> entry : data.entrySet()) {
             LocalDateTime dateTime = entry.getKey();
             Hour hour = new Hour(dateTime.getHour(), dateTime.getDayOfMonth(), dateTime.getMonthValue(), dateTime.getYear());
             series.add(hour, entry.getValue());
